refactor: move explorer layout into page slice
