+++ src/features/KeyboardShortcuts/KeyboardShortcuts.tsx
@@ -7,10 +7,10 @@
 import { useAtom, useAtomValue, useSetAtom } from 'jotai';
 import { useEffect, useRef } from 'react';
 import { useHotkeys } from 'react-hotkeys-hook';
-import { isSidebarOpenAtom } from '@/app/ui/AppSidebar/model/atoms';
 import { viewModeAtom } from '@/entities/AppView/model/atoms';
 import { useOpenFile } from '@/features/File/OpenFiles/lib/useOpenFile';
 import { searchModalOpenAtom } from '@/features/Search/UnifiedSearch/model/atoms';
+import { isSidebarOpenAtom } from '@/pages/explorer/widgets/ExplorerSidebar/model/atoms';
 import { activeTabIdAtom, openedTabsAtom } from '@/widgets/MainContents/model/atoms';
 
 const GLOBAL_HOTKEYS = {
