+++ src/pages/explorer/widgets/ExplorerSidebar/model/atoms.ts
@@ -1,0 +1,5 @@
+import { atom } from 'jotai';
+
+export const isSidebarOpenAtom = atom<boolean>(true);
+
+export const fileTreeModeAtom = atom<'all' | 'related'>('all');
