+++ src/pages/explorer/widgets/ExplorerSidebar/ui/ExplorerSidebar.tsx
@@ -1,43 +1,32 @@
-/**
- * AppSidebar - Sidebar container with FileExplorer
- * Provides resizable sidebar layout for file navigation
- */
-
 import { useAtom, useAtomValue } from 'jotai';
 import { ChevronDown, ChevronRight } from 'lucide-react';
 import type React from 'react';
 import { useRef, useState } from 'react';
 import { Sidebar } from '@/components/ide/Sidebar.tsx';
-import { viewModeAtom } from '@/entities/AppView/model/atoms';
-import { FileExplorer } from '../../../widgets/FileExplorer/FileExplorer.tsx';
-import { RelatedFilesView } from '../../../widgets/RelatedFilesView/RelatedFilesView.tsx';
-import { fileTreeModeAtom, isSidebarOpenAtom } from './model/atoms.ts';
+import { FileExplorer } from '@/widgets/FileExplorer/FileExplorer.tsx';
+import { RelatedFilesView } from '@/widgets/RelatedFilesView/RelatedFilesView.tsx';
+import { fileTreeModeAtom, isSidebarOpenAtom } from '../model/atoms.ts';
 
-export const AppSidebar: React.FC = () => {
+export const ExplorerSidebar: React.FC = () => {
   const isSidebarOpen = useAtomValue(isSidebarOpenAtom);
-  const _viewMode = useAtomValue(viewModeAtom);
   const [fileTreeMode, setFileTreeMode] = useAtom(fileTreeModeAtom);
   const containerRef = useRef<HTMLDivElement>(null);
-
-  // Collapsible states
   const [isFileExplorerCollapsed, setIsFileExplorerCollapsed] = useState(false);
 
   if (!isSidebarOpen) {
     return null;
   }
 
-  const projectLabel = 'Project';
-
   return (
     <div ref={containerRef} className="relative focus:outline-none">
       <Sidebar resizable defaultWidth={250} minWidth={200} maxWidth={800} className="h-full shadow-2xl">
-        {/* PROJECT */}
         <div className={!isFileExplorerCollapsed ? 'flex-1 flex flex-col overflow-hidden' : ''}>
           <button
+            type="button"
             onClick={() => setIsFileExplorerCollapsed(!isFileExplorerCollapsed)}
             className="flex w-full h-8 items-center justify-between border-b border-border-DEFAULT px-2 flex-shrink-0 hover:bg-bg-deep transition-colors"
           >
-            <span className="text-2xs font-medium text-text-tertiary normal-case">{projectLabel}</span>
+            <span className="text-2xs font-medium text-text-tertiary normal-case">Project</span>
             {isFileExplorerCollapsed ? (
               <ChevronRight className="w-3 h-3 text-text-muted" />
             ) : (
@@ -47,9 +36,9 @@ export const AppSidebar: React.FC = () => {
 
           {!isFileExplorerCollapsed && (
             <>
-              {/* Mode Tabs */}
               <div className="flex border-b border-border-DEFAULT">
                 <button
+                  type="button"
                   onClick={() => setFileTreeMode('all')}
                   className={`flex-1 px-2 py-1.5 text-2xs font-medium transition-colors ${
                     fileTreeMode === 'all'
@@ -60,6 +49,7 @@ export const AppSidebar: React.FC = () => {
                   All Files
                 </button>
                 <button
+                  type="button"
                   onClick={() => setFileTreeMode('related')}
                   className={`flex-1 px-2 py-1.5 text-2xs font-medium transition-colors ${
                     fileTreeMode === 'related'
@@ -83,5 +73,3 @@ export const AppSidebar: React.FC = () => {
     </div>
   );
 };
-
-export default AppSidebar;
